Update README.md (#61)
Added the video-per-video option to the readme as it is in the --help menu if you run in the command line but not here.

This actually really helped me as I was getting a "429: You have reached maximum request limit. Sleeping for 15 minutes" message if I tried to download all of the videos at once, and it wouldn't download anything. Using the video-by-video option seems to be working now.